src/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
tqdm==4.66.5

numpy==2.1.2
faiss-cpu==1.9.0

sentence-transformers==3.0.1
torch==2.4.1
